멤버변수와static의필요,data영역과code영역의존재
