+++ bin/twoconstructor/인스턴스.txt
