+++ src/Collapbration/객체 간의 협력.md
@@ -1,0 +1,7 @@
+\*\* 객체간의 협력
+
+> 객체 지향 프로그램에서 객체 간에는 협력이 이루어짐.
+> 협력을 위해서는 필요한 메세지를 전송하고 이를 처리하는 기능이 구현되어야함.
+> 매개 변수로 객체가 전달되는 경우가 발생.
+
+한 객체의 매개변수로 받아와서 처리하는 경우를 이야기하는듯하다.
