--- src/aboutConstructor/runHuman.java
@@ -1,9 +1,0 @@
-package aboutConstructor;
-
-public class runHuman {
-    public static void main(String[] args) {
-        Human tomas = new Human();
-        tomas.showInfo(80, 78, 37,"남","Tomas");
-    }
-    
-}
