+++ src/twoconstructor/runHuman.java
@@ -1,0 +1,9 @@
+package twoconstructor;
+
+public class runHuman {
+    public static void main(String[] args) {
+        Human tomas = new Human();
+        tomas.showInfo(80, 78, 37,"남","Tomas");
+    }
+    
+}
